+++ lib/super/reset.rb
@@ -4,6 +4,20 @@ module Super
   module Reset
     extend ActiveSupport::Concern
 
+    KEEP = {
+      # Defined by Rails
+      _layout: true,
+      _generate_paths_by_default: true,
+
+      # Defined in Super::SubstructureController
+      navigation: true,
+      page_title: true,
+
+      # Defined in Super::ApplicationController
+      current_action: true,
+      with_current_action: true,
+    }
+
     included do
       undef_method :index
       undef_method :show
@@ -14,10 +28,12 @@ module Super
       undef_method :destroy
 
       Super::SubstructureController.private_instance_methods(false).each do |imethod|
-        next if imethod == :_layout
-        next if imethod == :_generate_paths_by_default
-        next if imethod == :navigation
-        next if imethod == :page_title
+        next if KEEP.key?(imethod)
+        undef_method imethod
+      end
+
+      Super::ApplicationController.private_instance_methods(false).each do |imethod|
+        next if KEEP.key?(imethod)
         undef_method imethod
       end
     end