+++ app/controllers/super/substructure_controller.rb
@@ -231,48 +231,44 @@ module Super
 
     def index_view
       Super::Layout.new(
-        mains: [
-          Super::Panel.new(
-            Super::Partial.new("collection_header"),
-            :@display
-          ),
-        ],
-        asides: [
-          :@query_form,
-        ]
+        main: Super::ViewChain.new(
+          main_panel: Super::Panel.new,
+          main_header: Super::Partial.new("collection_header"),
+          main: :@display
+        ),
+        aside: Super::ViewChain.new(
+          main: :@query_form
+        )
       )
     end
 
     def show_view
       Super::Layout.new(
-        mains: [
-          Super::Panel.new(
-            Super::Partial.new("member_header"),
-            :@display
-          ),
-        ]
+        main: Super::ViewChain.new(
+          main_panel: Super::Panel.new,
+          main_header: Super::Partial.new("collection_header"),
+          main: :@display
+        )
       )
     end
 
     def new_view
       Super::Layout.new(
-        mains: [
-          Super::Panel.new(
-            Super::Partial.new("collection_header"),
-            :@form
-          ),
-        ]
+        main: Super::ViewChain.new(
+          main_panel: Super::Panel.new,
+          main_header: Super::Partial.new("collection_header"),
+          main: :@form
+        )
       )
     end
 
     def edit_view
       Super::Layout.new(
-        mains: [
-          Super::Panel.new(
-            Super::Partial.new("member_header"),
-            :@form
-          ),
-        ]
+        main: Super::ViewChain.new(
+          main_panel: Super::Panel.new,
+          main_header: Super::Partial.new("member_header"),
+          main: :@form
+        )
       )
     end