+++ test/lib/super/view_chain_test.rb
@@ -1,0 +1,6 @@
+require "test_helper"
+
+class ViewChainTest < ActiveSupport::TestCase
+  test "#erb" do
+  end
+end