+++ frontend/super-frontend/source/src/javascripts/application.js
@@ -1,4 +1,6 @@
+import "@webcomponents/custom-elements";
 import "details-element-polyfill";
+import "@github/details-dialog-element";
 
 import RailsUjs from "@rails/ujs";
 import { Application, Controller as StimulusController } from "stimulus";