+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.19.0] - 2021-12-13
+
 ### Improvements
 
 * `#member_actions` and `#collection_actions` can return an array of things that
@@ -388,7 +390,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Sprockets and Webpacker support. Defaults to Sprockets
 * Navigation bar with links to all admin controllers
 
-[Unreleased]: https://github.com/zachahn/super/compare/v0.18.0...HEAD
+[Unreleased]: https://github.com/zachahn/super/compare/v0.19.0...HEAD
+[0.19.0]: https://github.com/zachahn/super/compare/v0.18.0...v0.19.0
 [0.18.0]: https://github.com/zachahn/super/compare/v0.17.0...v0.18.0
 [0.17.0]: https://github.com/zachahn/super/compare/v0.0.16...v0.17.0
 [0.0.16]: https://github.com/zachahn/super/compare/v0.0.15...v0.0.16